requirements.txt
.
streamlit>=1.32
yfinance>=0.2
pandas>=2.0
numpy>=1.26
python-docx>=1.0
httpx>=0.27
pydantic>=2.0
loguru>=0.7
rich>=13.0
plotext>=5.3
typer>=0.12
fpdf2>=2.7
pypdf>=4.0

# Outreach automation (non-engine)
anthropic>=0.30
praw>=7.7
python-dotenv>=1.0
